feat: Add common response structure
