--- order-service/src/main/java/com/ed/orderservice/libs/common/ErrorResponse.java
@@ -1,20 +1,0 @@
-package com.ed.orderservice.libs.common;
-
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.Getter;
-
-@Getter
-public class ErrorResponse {
-	private final List<StackTraceElement> stackTraces;
-	private final String message;
-	private final HttpStatus status;
-
-	public ErrorResponse(List<StackTraceElement> stackTraces, String message, HttpStatus status) {
-		this.stackTraces = stackTraces;
-		this.message = message;
-		this.status = status;
-	}
-}
