+++ order-service/src/main/java/com/ed/orderservice/libs/common/HttpHeaderConstants.java
@@ -1,0 +1,11 @@
+package com.ed.orderservice.libs.common;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class HttpHeaderConstants {
+
+  public static final String HEADER_USER_ID = "X-User-Id";
+  
+}
