+++ order-service/src/main/java/com/ed/orderservice/libs/response/ApiResponse.java
@@ -1,0 +1,22 @@
+package com.ed.orderservice.libs.response;
+
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ApiResponse<T> {
+
+  private Boolean success;
+  private T body;
+  private LocalDateTime timestamp = LocalDateTime.now();
+
+  public ApiResponse(Boolean success, T body) {
+    this.success = success;
+    this.body = body;
+  }
+  
+}
