+++ order-service/src/main/java/com/ed/orderservice/libs/response/ResponseWrapper.java
@@ -1,0 +1,36 @@
+package com.ed.orderservice.libs.response;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class ResponseWrapper implements ResponseBodyAdvice<Object> {
+
+  @Override
+  public boolean supports(
+      MethodParameter returnType, Class<? extends HttpMessageConverter<?>> converterType) {
+    return true;
+  }
+
+  @Override
+  public Object beforeBodyWrite(Object body,
+      MethodParameter returnType,
+      MediaType selectedContentType,
+      Class<? extends HttpMessageConverter<?>> selectedConverterType,
+      ServerHttpRequest request,
+      ServerHttpResponse response) {
+
+    if (body instanceof ErrorResponse) {
+      return new ApiResponse<>(false, body);
+    }
+
+    return new ApiResponse<>(true, body);
+  }
+}
